requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
pydantic==2.8.2
langchain==0.2.14
aiohttp==3.10.5
python-dotenv==1.0.1
pytest==8.3.3
